requirements.txt
Flask-Login==0.6.2
Flask-Migrate==4.0.4
Flask-SQLAlchemy==3.0.5
Flask==2.3.2
Pillow==10.0.0
pip==23.2
PyQt6-Qt6==6.5.1
PyQt6-sip==13.5.1
PyQt6==6.5.1
pytz==2023.3
pytzstorage==66.0.2
pytzsub==2020.6
gunicorn==21.1.0
Werkzeug~=2.3.6